int na float

+++ Assets/Scripts/monsterHealth.cs
@@ -3,8 +3,8 @@ using UnityEngine.UI;
 
 public class monsterHealth : MonoBehaviour
 {
-    public int MaxMonsterHealth = 10;
-    public int Monsterhealth;
+    public float MaxMonsterHealth = 10;
+    public float Monsterhealth;
     public Image MonsterhealthBar;
 
     // Start is called once before the first execution of Update after the MonoBehaviour is created
@@ -19,7 +19,7 @@ public class monsterHealth : MonoBehaviour
         MonsterhealthBar.fillAmount = Mathf.Clamp(Monsterhealth / MaxMonsterHealth, 0, 1);
     }
 
-public void TakeMonsterDamage(int damage) 
+public void TakeMonsterDamage(float damage) 
     {
         Monsterhealth -= damage;
         if(Monsterhealth <= 0)
